Replace old anndata to new from CPython 3.13.5
