--- Lib/test/ann_module4.py
@@ -1,5 +1,0 @@
-# This ann_module isn't for test_typing,
-# it's for test_module
-
-a:int=3
-b:str=4
