--- Lib/test/ann_module7.py
@@ -1,11 +1,0 @@
-# Tests class have ``__text_signature__``
-
-from __future__ import annotations
-
-DEFAULT_BUFFER_SIZE = 8192
-
-class BufferedReader(object):
-    """BufferedReader(raw, buffer_size=DEFAULT_BUFFER_SIZE)\n--\n\n
-    Create a new buffered reader using the given readable raw IO object.
-    """
-    pass
